backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
mysqlclient>=2.2
python-dotenv>=1.0
django-filter>=23.5
gunicorn>=21.0
